Time: 0 ms (100%), Space: 32.1 MB (45.76%) - LeetHub

+++ 1055-pairs-of-songs-with-total-durations-divisible-by-60/1055-pairs-of-songs-with-total-durations-divisible-by-60.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
-    long long int numPairsDivisibleBy60(vector<int>& time) {
+    long int numPairsDivisibleBy60(vector<int>& time) {
         unordered_map<int, int> mp;
-        long long int res = 0;
+        long int res = 0;
         for(int t: time) mp[t % 60]++;
         if(mp.count(0)) res += (1LL * mp[0] * (mp[0] - 1)) / 2; 
         if(mp.count(30)) res += (mp[30] * (mp[30] - 1)) / 2;
